refactor: extract ai settings helpers
